LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-events-2.optional-expected.txt

one select0-button
          one two    select7-button

FAIL Button controller code should not run if the mousedown event is preventDefaulted. assert_false: Listbox shouldn't have opened yet expected false got true
PASS Listbox controller code should not run if the mousedown event is preventDefaulted.
FAIL <select> should fire input and change events when new option is selected. assert_false: expected false got true
PASS <select> should not fire input and change events when new selected option has the same value as the old.
FAIL <select> should fire input and change events when option in listbox is clicked assert_equals: input event shouldn't fire when selected option didn't change expected 1 but got 0
FAIL Check that <Space> opens <select>. assert_false: <Enter> should close select expected false got true
FAIL Check that <Space> opens <select> when <select> specifies tabindex assert_false: <Enter> should close select expected false got true
